requirements.txt
tencentcloud-sdk-python~=3.0.203
